app/files/serializers.py
from rest_framework import serializers
from files.models import File


class FileSerializer(serializers.ModelSerializer):

    class Meta:
        model = File
        fields = '__all__'

    def to_representation(self, instance):
        data = super().to_representation(instance)
        request = self.context.get('request')

        if request and request.method == 'POST':
            data = {'file': data['file']}

        return data
